adds Login Page [JSX]
COMMENTS:
Adds the Login Page to the React Twitter App.

Links the Public and Login Pages.

Uses a temporary user account to test the login validation code.

Encapsulates the user credentials in a object literal, as jestrade
suggested in his React lecture, instead of storing them in separate
`username' and `password' placeholders.