adds Home Page [JSX]
COMMENTS:
The Home Page contains a `welcome' message and a `logout' button,
for it is used for testing the routing of (un)authenticated users.

Routes new users to the Login Page upon successful sign up.

Routes authenticated users to the Home Page.

Routes unauthenticated users to the Public Page.

User credentials are saved in the local storage (for now).