authenticates with the API for user login
COMMENTS:
Disables temporarily the Signup Component, for the API has yet to
support user account creation (or signup).

Gets the Twitter API URI from process environment variables.

Sends a POST request to the Twitter API with the credentials input
by the user. Saves the JSONWebToken JWT in the local storage upon
success for future authentications; otherwise informs the user
that the supplied credentials are invalid.

Routes authenticated users to the Private component /home.

The Private component now checks if the token is present in the
local storage for authorization.

Upon successful login the Private component saves the logged-in
status in the local storage just in case the user attempts to
return to the login page manually by changing the route. If that's
the case the Login component routes the user back to /home. Users
must logout first if they wish to switch accounts.